pyproject.toml
[tool.poetry]
name = "drf-todo"
version = "0.1.0"
description = ""
authors = ["mogendi <mogendi.mongare@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
